tast: add arc_p version of multivm.login test
Existing test multivm.login has no history of running on
ARC++ (container). Adding the new flavor, so that it buildst this
history going forward, and it gets ready to issue data for
compating ARC++ and ARCVM on upcoming models we release.
Also created a short timeout for connecting to DBUS, to fail
faster if running against older system, or if there's a regression
in the connection to resourced.

BUG=b:217276698

TEST=tast -verbose run __dut__ multivm.Login.arc_p

Change-Id: I0e88de20e74fc54fc7959b91e67c5e23bec77b8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3494197
Tested-by: Alexandre Marciano Gimenez <raging@google.com>
Auto-Submit: Alexandre Marciano Gimenez <raging@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Alexandre Marciano Gimenez <raging@google.com>