tast-tests: Add browserfixt.CreateWindows for migration
CreateWindows is a util that makes the transition from
ash.CreateWindows easy for both ash-chrome and lacros-chrome.
It is necessary to address the difference of application life cycle
between them.
Unlike ash-chrome which instance is always available even with
no window open, lacros-chrome needs to opens at least one extra blank
window before instantiating.

Also add example.Browser to test it out.

By the way, this works only with a blank URL, but not others because
of lacros.LaunchWithURL failing to open it on browser startup.

See go/lacros-tast-porting for the porting guidelines.

BUG=None (TBD)
TEST=tast run example.Browser*

Change-Id: I8d3d64312fb8ac0123c90f4b461ab85f74fe2ddd