[tast] Activities start with default intent action
The C2E2ETest actvity expects an intent action with all intents sent to
it otherwise it will crash for null pointer dereference. A refactor to
the act.Start API now passes the "-n" flag before the intent component
name always, but when the "-n" flag is passed, the default intent action
of android.intent.action.MAIN is not passed. This change will allow the
"-n" flag to remain in place while always passing an intent action.

The tast test arc.VideoDecodeAccelPerf.* was failing because it uses the
above APK.

BUG=b:203214749
TEST=tast run arc.VideoDecodeAccelPerf.h264_1080p_30fps_vm on eve DUT

Change-Id: Iaa190e9da1e466951521cfb46c054df7a084e3f0