No dex2oat check for arc.PowerIdlePerf.noarc
The noarc version of the test is failing because we try to check that
dex2oat is not running, but arc is not running, so the check fails.

This change only does the check if arc is running.

BUG=b:220568784
TEST=tast run dut arc.PowerIdlePerf.noarc arc.PowerIdlePerf.vm

Change-Id: Ie58717b907c1ab21f641ec6f5bce548165a813b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3520771
Tested-by: Charles William Dick <cwd@google.com>
Auto-Submit: Charles William Dick <cwd@google.com>
Reviewed-by: Eliot Courtney <edcourtney@chromium.org>
Tested-by: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: Hikaru Nishida <hikalium@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Charles William Dick <cwd@google.com>