Crostini: Add Screendiffing to this test.
Note that even though `crostiniBusterLargeContainer` and
`crostini.CrostiniAppTest` HardwareDep is registered with
this test, on many test runs, the container doesn't
appear to have Gedit pre-installed, resulting in test
failures in the "open temp file with Gedit" step. I have
added a check to ensure Gedit exists, or installs, prior
to the test progressing.

BUG=b:220779813
TEST=Ran on hatch (Akemi) DUT and remotely on lab Atlas boards.

Change-Id: I39c1f9a960d6881b9c4d29843807aafa99d287ca