tast-tests: Update Lacros metadata to settings*
The following tests need a lacros variant but the feature is not yet
completed. See https://crbug.com/1306270.
LacrosVariantNeeded to settings.ChromeOSPageInfo
LacrosVariantNeeded to settings.TermsLinkClickable

The following test doesn't need a lacros variant as it doesn't interact
with a browser:
LacrosVariantUnneeded to settings.SearchSections

See go/lacros-tast-porting for the porting guidelines.

BUG=b:218523908
TEST=fast_build.sh -T

Change-Id: I73f440b349f66eb893edb054aab243b1ce5ca880