tast-tests: Update Lacros metadata to shelf*
The following tests need a lacros variant, but the migration is
blocked on lack of the Test API support to enable Lacros via chrome.New.
LacrosVariantNeeded to shelf.AppRearrangement

The following test doesn't need a lacros variant as it doesn't interact
with a browser:
LacrosVariantUnneeded to shelf.MenuOverflow

See go/lacros-tast-porting for the porting guidelines.

BUG=b:219815913
TEST=fast_build.sh -T

Change-Id: I8ce2b5228942f42c039e9140407b8109297cc191