Properly close gRPC clients in nearby share remote fixture
The last few remote tests that run are constantly failing due to SSH
connection failures. The issue appears to be that we are reaching the
limit for the # of open SSH connections on the DUT. This is caused by
leaving gRPC client connections open between fixtures, so we need to
properly close those when we are done with them.

BUG=b:220942040
TEST=tast run -companiondut=cd1:<dut1> <dut2> '("group:nearby-share-remote")'

Change-Id: I35f0a1bada2e908bf31361d93868527fd38ee9be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3522757
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Auto-Submit: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Kshitij Pancholi <panchok@google.com>