tast-tests: wifi: improve MBOAssocDisallow
There was a timing issue in this test.  Since we can't configure hostap
with 'mbo_assoc_disallow' (as this is not a config parameter) we have
following scenario possible:
- AP is configured
- shill/WPA starts scan
- we change the 'mbo_assoc_disallow' via CLI
- report comes but the results are from before our modification
- then we try to connect
- connection fails but in the PCAP we see the association request and we
  fail the test

In order to prevent this scenario let's flush the scan cache before
trying to connect.  This way the AP should get rediscovered with
'Association Disallowed' attribute and we can safely check absence of
association request in subsequent connection attempt.

BUG=b:222580782
TEST=Manual.  Find in stainless results a board/model that fails this
  test (at least some times).  Run it multiple times without this change
  to reproduce and then try the same with this change.  Confirmed this
  fix on cherry/tomato.

Change-Id: Ibcc6bd1af6635be854c975e51687b9f385874f33