tast-tests: Update Lacros metadata to familylink*
The following tests seem to need a lacros variant since it interacts
with a browser. However, the feature may not be completed yet, so
migration should be planned separately. See crbug.com/1206315 for
progress.
- familylink.DeniedSitesBlocked
- familylink.MultipleSignInDisabled
- familylink.PolicyLogin  (testing custom fixture)
- familylink.UnicornLogin  (testing custom fixture)

The following test doesn't need a lacros variant as it doesn't interact
with a browser:
- familylink.AddPersonFlow
- familylink.BedTimeLimit
- familylink.ChildWallpaperSync
- familylink.DailyTimeLimit
- familylink.EducoexistenceInsession

See go/lacros-tast-porting for the porting guidelines.

BUG=b:224838424
TEST=fast_build.sh -T

Change-Id: I55bbb859b22e673caba8b28b6a17f2420469184b