[Tast] Fix flakiness of network.CellularPolicyConnection test
This test failed on dedede device in lab, because sometimes when
clicking on connect button in cellular detail page, it may get back to
to the internet subpage. This CL removes the step of connecting to
each eSIM network before performing actual tests since we don't need it
to ensure Shill config get created anymore.

BUG=b:221466540
TEST=tast run ${DUT} network.CellularPolicyConnection

Change-Id: Iff94c597df0269fa08116a394c3fa84aade6a358