tast-tests: refactored Router.RouterTypeName() to RouterType.String()
This also returns a different value for the AX router type, "AX",
than what ax.RouterTypeName() provided, "AsusAX". The former is
used as it is more general and the latter and the usage thus far
is really only for logging errors so changing it has no impact.

BUG=b:211517577
TEST=tast run -var=router=$LEGACY_ROUTER -var=pcap=$PCAP $DUT \
wifi.SimpleConnect.80211a

Change-Id: Ibfc15c1a00904138c28530f564704b23c9f61d6e