factory: Add toolkit service for installation/uninstallation
The commands for installing/uninstalling factory toolkit in remote and
local tests are the same, but cannot share between them as the API for
sending shell command to the DUT is different in remote and local tests.

Create a dedicated service for handling requests from remote tests to
install/uninstall factory toolkit so that we only need to maintain one
implementation for instructing DUT to install/uninstall the factory
toolkit in local test packages.

BUG=b:225083013
TEST=None

Change-Id: I2e2fa0b085b8231e61e6f1dae5d3cb397ab77e3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3531253
Reviewed-by: Meng-Huan Yu <menghuan@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>