factory: Add toolkit service for installation/uninstallation
The commands for installing factory toolkit in remote and local test are
the same, but cannot share between them as the API for interacting with
the DUT is different.

Create a dedicated service for handling requests from remote tests to
install/uninstall factory toolkit so that we can keep and only need to
maintain one implementation in local packages.

BUG=b:225083013
TEST=None

Change-Id: I2e2fa0b085b8231e61e6f1dae5d3cb397ab77e3e