Add OS install test
This test expects to be run against a DUT booted from a reven
installer. The test will start the install process through the UI, wait
for it to succeed, then power off the device. It will then wait for the
device to boot back up into the installed system and do some basic
checks to verify that install was successful.

Steps for testing with QEMU:
1. Make a destination disk: `qemu-img create -f raw installed.bin 16G`
2. Launch qemu with the reven installer, and `-hdb installed.bin`
3. Run `tast run vm install.OsInstall`
4. Wait for the VM to power off
5. Launch qemu again, this time with `installed.bin`
6. Wait for the test to succeed.

Can also test that it detects a failure to boot into the installed
system by changing step 5 to boot back into the reven installer, and
observe that the test then fails.

BUG=b:187529021
TEST=See above

Cq-Depend: chromium:3587282
Change-Id: Id6109e15d9d74b227853ee150cb2c79cdae4560e