Revert "hwsec: restart ui before unmount cryptohome"
This reverts commit 93f77ed9bf0337f80aebe9192747785341c2f931.

This change is causing WiFi tests that interact with the certificate
store for auth to fail.  Not having these tests is a release blocker.
Reverting this change is the fastest way to get test coverage restored.

BUG=b:224802056
TEST=Checked out a dut, ran affected tests with the changes to restart
the UI, and they fail.  Reran the tests with the UI restart
functionality reverted, and they pass.

Change-Id: I90f2ee94776ff290ed5fcf4b9dd32e438c13ea5a