u2fd: Promote WebauthnUsingPassword.tpm1 to CQ
Remove the informational attribute from the u2fd.WebauthnUsingPassword
test (tpm1) so it runs on CQ. This test passed 17 consecutive builds after
landed until it meets an issue that is fixed by now, and it continues to
pass. There's no observed flakiness in this test, making it suitable for
running in CQ. In addition, we currently have no tests in CQ about u2fd
such that CQ will pass even if u2fd fails to start at all. We want a
stable test about u2f/WebAuthn running in CQ to monitor if any CL breaks
its behavior.

However, the WebauthnUsingPassword.gsc seems to be a little flaky (though
not from the test itself). Although it would be great to have them both
in CQ to make sure WebAuthn is successful on both gsc and TPM 1.2
devices, we might want to wait until it's more stable.

BUG=None
TEST=tastboard consistently passing

Change-Id: I06fb983614c5bd074536f733647a1c10221867b7