u2fd: Wait until the pp prompt to appear before pressing
Fix a bug that we pressed the power button too fast, before u2fd even
sends the request to cr50 and prompts for the press. This will result in
the power button pressed before power manager is told to ignore it, and
the screen or device would be closed.

Wait for the prompt's UI to appear before pressing the power button.

BUG=b:213852189
TEST=tast run -var=servo=$SERVO $DUT hwsec.WebauthnU2fMode

Change-Id: I222d74a2aec5f237095499686f890e1510ffc74b