faft: Add short delay before enabling USB stick
It seems that if you disable the USB flash drive, and then quickly
enable it again, then when you boot to recovery mode, it will boot, and
start to work, but once the ethernet connects, it will crash.

Adding a short delay before enabling the usb flash drive seems to fix the
problem.

BUG=b:225178570
TEST=tast run all faft_smoke tests on eldrid

Change-Id: Ib03e2ba6bfba7bb40dcf3195aeb1816e9033fc83