Add prod and dev versions of a few nearby tests
Add some fixtures and a few subtests for Nearby Share that use the dev
and prod versions of Nearby Share on the Android device. The active user
account on the Android determines which version of Nearby it uses, so
the only new configuration that's needed is fixtures with different
Android logins.

BUG=b:207026818
TEST=tast run <dut> '("group:nearby-share-dev")'
TEST=tast run <dut> '("group:nearby-share-prod")'

Cq-Depend: chrome-internal:4623096, chromium:3534343
Change-Id: Ia68e140cb328a0121dfea697ce59a8448b19ee91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3534445
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>