soc_properties: informational test to validate ro.soc.*
Verifies ro.soc.manufacturer and ro.soc.model conform to the expected
format. This is informational since the properties do not work for all
SOC models of possible devices. It will be switched to critical when all
models can be detected.

The ro.soc.model regexp is based on that found in
go/cdd-proposal-signoff, but made a little more restrictive.

BUG=b:175610620,b:225373614
TEST=Verified a more restrictive regexp for either prop causes test failure
TEST=Verified a missing property causes test failure

Change-Id: Id7a6a3217c46f07754d6aca5113051934e3ea2d2