policy: Fix remaining flakiness of FileSystemWriteBlockedForUrls test
The 'Save File as' file picker can only be triggered from JavaScript in
response to a user gesture. It appears that this test was flaky because
Chrome would sometimes not detect the click executed by Tast as a user
gesture.

This CL fixes this by re-trying the click if such an error occurrs. The
test now also logs the JavaScript error message for ease of future
debugging.

BUG=chromium:1297527
TEST=tast -verbose run <IP> policy.FileSystemWriteBlockedForUrls*

Change-Id: Ib9a716f3252a96dd7cd674414954b73cd66e82ba