tast-tests: Ignore command verification in PerfETM
Apparently there is a bug in perf inject which treats a traced command
as "perf" in the timeless decoding. Similar problem - b/163171518.
It's not a blocker for devices in the field since we are going to trace
only commands which are already running. Those are not affected.

Ignore the command verification error to unblock ETM testing.

BUG=b:225407997
TEST=tast run $dut platform.PerfETM

Change-Id: Ib4ce0c8466301c695ff036f9bb5accda38a1aac3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3536476
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Tested-by: Denis Nikitin <denik@chromium.org>
Commit-Queue: Denis Nikitin <denik@chromium.org>