chrome/internal: Set --force-color-profile and --force-raster-color-profile
These flags force chrome to always render in the sRGB color space. Tast
tests are not color-space aware (and can't easily be made so), but need
to be able to check the rendered colors in screenshots. Forcing chrome
to always render in a fixed color-space, rather then the display's
color-space, makes this work without a lot of work.

Note that --force-color-profile is currently ignored on ChromeOS, so
this CL also requires CL:3563607 to be effective.

BUG=b:221643955
BUG=b:221695443
TEST=Ran screenshot tests

Change-Id: Ieaa9ee961b7bbe4d42b01a7eb19b72044fb824c4