drivefs: Add manual test to renew refresh tokens
The DriveFS tests now use pooled GAIA accounts instead of the single
account. These accounts have a refresh token assigned to them, however,
if one or more expire it is best to renew all of them at once. This is a
tedious manual process so introduce a tast test that can log all the
refresh tokens at the end. Deliberately keep this manually invoked (not
running informational or mainline) to avoid leaking the refresh tokens.

BUG=chromium:1299712
TEST=tast run filemanager.DrivefsRenewRefreshTokens

Change-Id: I6dd520a6240e76c195e5282848de49044eaa27ad