tast-tests: Dump bugreport on package install failure
There are cases where a scheduled job in finsky doesn't execute. To file
a bug with the managed play team, we need a bug report. Since the bug is
not easily reproducible on manual runs, adding the bug report dumping
logic in the wait for package install helper method.

BUG=b:186274474
TEST=tast run DUT enterprise.ARCProvisioning.vm

Change-Id: I14a804e972af0b7cca3f14bbb488265663b17a8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3539316
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>