tast-tests: Update AxSimpleConnect subtest names to use broadcast/hidden
The original subtest names in AxSimpleConnect were confusing as "open"
could refer to a network without a password (wpa-psk) or a network that can be
discovered (as opposed to hidden network). Examples of confusing names
would be '80211axopenhidden'.

One naming option would be to only specify when a test is "hidden" and
not specify the "not hidden" feature in the test name if the test is not
hidden. However, this still makes the test name, 80211axopenhidden, a
valid name and still confusing.

Instead, instead of having the notion of a open/hidden network, we have
the option of having a broadcast/hidden network. 80211axopenhidden still
is a valid name, but other test names that aren't hidden would display
"broadcast" in the names (i.e 80211axwpabroadcast6ghz).

BUG=None
TEST=tast run wifi.AxSimpleConnect

Change-Id: I36ccbce59568f198797d5e01deab2c57bca23679