tast-tests: Update AxSimpleConnect subtest names to use broadcast/hidden
The original subtest names in AxSimpleConnect were confusing as "open"
could refer to a network without a password (wpa-psk) or a network that can be
discovered (as opposed to hidden network). Examples of confusing names
would be '80211axopenhidden'.

We only specify the hidden or wpa attributes in the test names. We
disambiguate "open" networks to be "broadcast" if it's not hidden and
"open" if the network does not require a preshared key. Open and
broadcast are only present in the test description, but are otherwise
omitted from the subtest names.

This change only seeks to update to update the test names and test
descriptions.

BUG=None
TEST=tast run wifi.AxSimpleConnect

Change-Id: I36ccbce59568f198797d5e01deab2c57bca23679
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3541104
Reviewed-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Tested-by: Billy Zhao <billyzhao@chromium.org>
Auto-Submit: Billy Zhao <billyzhao@chromium.org>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>