crostini: enable fio perf test.
The test was disabled long time ago because of flakiness, but it is
impossible to find the reason of the flakiness cuz all the logs are
gone. This CL re-enables nightly crosbolt tests for crostini to see if
it is still failing and subsequently address the failure reasons to
provide a stable crosvm disk IO perf test framework.

BUG=b:226198367
TEST=tast --verbose run <DUT> crostini.DiskIOPerf

Change-Id: I7c60b792709d721f5a1f6e9784bae5e7dde383a0