wifi: Refactor WiFi Device Info bundle after moving
Refactor WiFi Device Info bundle leftovers after moving the most
of the WiFi device info to tast for sharing with hwdeps.
Most of its code was moved to avoid code duplication.
LogBandwidthSupport() wasn't moved, as it was not needed in tast
and was disturbing dependencies (causing loops). It has been inlined
instead in the only place it was used.

Refactor three tests using device info so it uses now the tast's version.

BUG=b:218337813
BUG=b:187699768
TEST=wifi.Caps, wifi.Driver, wifi.Regulatory

Cq-Depend: chromium:3489491
Change-Id: I173a1fd7f19cea79add27ba83f7b4e7a29a9a838