login: Set timeout for ShowPassword test
The tast.login.ShowPassword test (in both PIN and password variants) is
flaky before this commit because of exceeded deadlines, with pass rates
at 88%.
Before this commit, the test doesn't specify a timeout, and the default
timeout is evidently too low. This commit sets an explicit timeout so as
to reduce flakiness. The commit also sets the LacrosStatus to satisfy
the linter.

BUG=chromium:1309044
TEST=tast -verbose run $DUT login.ShowPassword.*

Change-Id: I3363438e3da33b61c1cf23580e50a01c3b14bae2