drivefs: Enable drivefs specific flags at fixture level
DriveFS has it's own concept of flags, these are sometimes decoupled
from the ChromeOS specific flag. To ensure features are appropriately on
the right combination of flags, expose the drivefs flags via a fixture
parameter that can supplied.

This uses the command_line_args file to pass through the arguments then
restart DriveFS. This is used in the filemanager.DrivefsDssOffline test
to ensure the 2x switchblade flags are enabled on startup, which appears
to be one of the reasons for the test currently failing.

BUG=chromium:1299712
TEST=tast run filemanager.DrivefsDssOffline

Change-Id: Ia4a16eb8e0868206be752497ebbd38161376c4c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3544849
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>