u2fd: Check operation result without looking at log
Currently the WebAuthn tast tests parses the chrome log and see if there
is log indicating a successful attempt of Make Credential or Get
Assertion. But parsing chrome log takes a lot of time and we cannot
control how much log there is when running this test. Check suitable JS
attributes on the site to determine the status instead of checking the
log.

BUG=chromium:1309146
TEST=run u2fd.* tests and hwsec.WebauthnU2fMode

Change-Id: I491ae8e04ca2c5ea1fa2329fc1150b6397fa4617
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3544853
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: David Stevens <stevensd@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>