power.SuspendBattery: Improve reliability of checking battery charge
Currently this test does not attempt to retry if there is a problem
communicating with the servo while attempting to check the battery
level. This commit changes it so that these communication attempts are
tried up to three times to account for any transient communication
issues.

BUG=b:221888911
TEST=Ran on zork device and verified that checking the battery level
continues to work

Change-Id: I2eb6df2ec9a5fb78d0ec6b1dee7a377cdc2acafc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3549898
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Robert Zieba <robertzieba@google.com>
Commit-Queue: Robert Zieba <robertzieba@google.com>