power.SuspendBattery: Make determining sleep state support more robust
Currently this test attempts to determine which sleep states are
supported by reading from `/sys/power/state` and `/sys/power/mem_sleep`.
However this can be unreliable as a DUT may have the proper strings in
those file but be unable to enter those states. This commit makes it so
that the test attempts to enter those states to fully determine if they
are supported and produces specific errors for when they aren't

BUG=b:221888911
TEST=Ran on zork and nipperkin devices, verifying that sleep states are
detected appropriately

Change-Id: I78dc38c5b634975becad8d72618b4caf812207c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3549899
Tested-by: Robert Zieba <robertzieba@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Robert Zieba <robertzieba@google.com>