factory: Add HWID database test
The HWID database generation is composed of the probing and then
building a database. The probing stage could be affected by the command
, tools it calls, or different hardware types in the real world. The
probing output is directly passed to building stage, thus any
unexpected output or failure breaks building the database. Add a test to
ensure the flow from probing to building database is valid, and catch
any possible regression that breaks this flow in lab.

BUG=b:223075358
TEST=tast run $DUT factory.HWIDDatabase

Change-Id: I748711017f6d8c9918d03439cdf1aadd2801dcc2