memory: add ManaTEE memory metrics
Add ManaTEE stats to memory metrics. The primary metric is the amount
of memory used by the hypervisor outside of directly backing guest
memory. This also adds metrics for total memory visible in the
hypervisor, which tracks memory allocated very early in the hypervisor
kernel's boot, and unevictable memory, which tracks initramfs size.

BUG=b:224657537
TEST=multivm.Login.novm

Change-Id: I08e4da093d88c9e8d1bf1a900f86655050afa207