Change the order of successful payment check
In Change-Id Icedbf92ac008c6734f42efc17d7e8e2a98734b84 we introduced a possibility to have optional require auth screen. Here's the way it works: we look for required auth screen. If it exists, we do the required auth flow. If it doesn't we do nothing. However, it takes time to identify, the screen does not exist. A successful payment screen is present only for a short period of time. As result, if there is no required auth screen, we could be waiting for it, while successful payment screen appears and disappears. And only then we check for the payment screen, which is gone by now.

Change the order of the checks. Check for a presence of successful payment screen before checking for required auth screen. This is possible, because both screens are rendered by android at the same time.

BUG=b:226672330
TEST=tast run dut appsplatform.CheckSkuPurchase.vm

Change-Id: I8d072e07f9a6f67275798a60e13e1d48e7cd7790