Crostini: Fix failing test that depends on a chrome extension installation.
Test was originally written as part of b/217465949. However, it was failing consistently
on eve and atlas-kernel next (please see the bug for this CL for details).

This failure is because the context menu option for the default text app
which is a chrome extension was not showing showing as expected, but only on
the first run of the test.  Subsequent runs generally automatically fixed this
issue with no change to code.  Turns out this is likely because the installation
of the text app chrome extension wasn't being registered and picked up (not within
a 10-minute timeout,anyway).

This change adds a check for the installation with
`ash.WaitForChromeAppInstalled`.

BUG=b:226240873
TEST=Ran on lab eve and atlas boards.

Change-Id: I3e16d64ecc3a10073297a7c0b9d25294995e349c