tast-tests: deflake glideTyping on betty
The glideTyping test is flaky on betty board. After investigation, the
word is correctly shown in the textField, but the test failed to find it
in the UI tree because of high latency on betty.

Increasing the number of retries may solve the issue.

BUG=b:223081262
TEST=tast run <DUT> inputs.VirtualKeyboardGlideTyping

Change-Id: I51aeed52a96ca82048ecbbc2bdac63a187cf4a79
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3551127
Tested-by: Alvin Jia <alvinjia@google.com>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Commit-Queue: Alvin Jia <alvinjia@google.com>