tast-tests: add fine-grained security test
With introduction of more fine-grained security for WiFi networks we
split broad PSK SecurityClass into several security settings.  In order
to test that functionality wifi.FgsecConnect test is introduced.  Also
"psk" and "802_1x" security constants are renamed to reflect that they
now stand only for SecurityClass.

In the initial phase of fine-grained security deployment we don't plan
on changing the behaviour, so we will support all WPA modes in a single
service.  This type of Security is called "wpa-all".  The more detailed
information is kept internally and used during metrics collection.

BUG=b:226138205
TEST=tast run ... wifi.FgsecConnect.*

Change-Id: If18a13008fe71dd239b509aec47eae000df6df54