tast-tests: better cleanup for wifi.PersistenceBluetoothSansWifi
It has been observed that on some platforms (f.e. zork/gumboz)
wifi.PersistenceBluetoothSansWifi test causes wrong (re)initialization
of the wlan device.  In particular randomization support gets turned
off.  Correct settings for that model is:

    MACAddressRandomizationEnabled = True
    MACAddressRandomizationSupported = True

but f.e. out of 10 runs of this test I ended up 8 times with:
    MACAddressRandomizationEnabled = False
    MACAddressRandomizationSupported = False
and 2 times with
    MACAddressRandomizationEnabled = False
    MACAddressRandomizationSupported = True

This suggest that there is some problem with reinitialization of that
device once we reboot with wifi disabled.  This CL proposes workaround
for that (until b/227582293 is resolved) by adding additional reboot
after reenabling wifi at the end of the test.

BUG=b:219083050, b:227582293
TEST=Manual.  Run this test multiple times on zork/gumboz and check that
     after each run we end up with randomization both enabled and
     supported.

Change-Id: Icb95471133650e0690fb4b7253754e0776aaa12f