hwsec: End test when one subtest fails in WebauthnU2fMode
The different subtests in WebauthnU2fMode are not independent as a
failed state of chrome browser or u2fd will affect the next subtest. The
only useful information from the logs when the test faile is the first
failure. Current behavior produces too noisy error message and also
provides a screenshot at the end of the test that is not useful.

BUG=b:213852189
TEST=tast run $DUT hwsec.WebauthnU2fMode

Change-Id: I3817a0d36a2089b2a18617e1556c3a557871ba8b