Fix how launcher.OpenExpandedView detects bubble launcher
OpenExpandedView waits for launcher animations in different way
depending on whether the launcher is fullscreen or "bubble" launcher.
For fullscreen launcher, it uses a dedicated API, while for bubble
launcher it uses uiauto package. This flips how the method finds out
which method to use - instead of waiting for bubble launcher node to
show up, and then falling back to waiting for fullscreen launcher state,
it now first attempts to wait for the fullscreen launcher state, and
falls back to uiauto API on failure that indicates the launcher is
expected to be shown in bubble UI.
The later approach avoids multisecond timeout when checking bubble
launcher node existence, so it should be more efficient and less
susceptible to flakes.

BUG=chromium:1310750
TEST=launcher.SearchBuiltInApps* on kevin

Change-Id: I233e3cb8a57771dea501d1a3faa2f9810e500b94