[ZipMount] Stop relying on mounted archives order
When mounting a selection of multiple encrypted archives, cros
disks returns a signal for each of these archives indicating
that a password is needed. However, we can't predict the order
in which these signal are going to be received.

Modify the tests to avoid relying on the archive mounting order.

As described in crbug.com/1308370, on password failure, the old
password is not selected anymore. Fixing the test by manually selecting
the old password (ctrl+A).

BUG=chromium:1308378, chromium:1308370
TEST=tast run $DUT filemanager.ZipMount.*

Change-Id: Ia3014a3046390e86df3fdc5294a58d821b201a9f