[ZipMount] Stop relying on mounted archives order
When mounting a selection of multiple encrypted archives, cros
disks returns a signal for each of these archives indicating
that a password is needed. However, we can't predict the order
in which these signal are going to be received.

Modify the tests to avoid relying on the archive mounting order.

BUG=chromium:1308378
TEST=tast run $DUT filemanager.ZipMount.*

Change-Id: Ia3014a3046390e86df3fdc5294a58d821b201a9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3555512
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: David Stevens <stevensd@chromium.org>
Commit-Queue: Jeremie Boulic <jboulic@chromium.org>
Tested-by: Jeremie Boulic <jboulic@chromium.org>