u2fd: Fix autosubmit behavior and disable it by default
Because of this bug (http://b/206379831), chrome image with DCHECK
enabled will crash when PIN autosubmit is enabled. This results in much
inconvenience when running this test on chrome image built from source.
It doesn't really matter whether autosubmit is enabled in this test, so
disable it by default.

When testing, I discovered that some logic that handles disabled
autosubmit is incorrect (and not caught because previously it's always
enabled). Fix the logic so that the test is correct whether autosubmit
is enabled or disabled.

BUG=b:206379831
TEST=tast run $DUT u2fd.WebauthnUsingPIN

Change-Id: I3772884e57604e85fca18144e3c9c9aadd83c830
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3555516
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>