tast-tests: Add PEAP wifi to rollback test
Extend rollback_service to support the verification of multiple
networks and make these transparent to the tests that use the
service. Add verification of preservation of a PEAP wifi.

Rollback integration tests only check a PSK network is preserved,
but more type of networks are supported by rollback. Extend the
service so it is possible to add more type of networks to the tests.
The configuration of the supported networks is handled in the service
and transparent to the tests that make use of it.

BUG=chromium:1264389, b/217546714, b/227562233
TEST=tast -verbose run DUT policy.EnterpriseRollbackInPlace
tast run -var=updateutil.tlwAddress="localhost:7151" localhost:2200
autoupdate.EnterpriseRollbackPreviousVersion.* all
All the tests pass when PEAP configuration is not included (current
state of the tests with just PSK still works with the modification).
Tests pass when also adding the new PEAP wifi configuration.

Change-Id: I33f5601165d546caf9ed9265365f7f9abd0db3ac