tast: Add test for WiFi concurrency capabilities
Add a test to record the WiFi concurrency capabilites of devices, as
advertized by the `iw list` command. Currently, this test only logs the
devices advertized support, meaning (1) it does not assert any required
support nor does it (2) validate that the device actually provides its
claimed support.

For (1), the intent is for this test to be updated in the future once
concurrency support has been added to AVL to fail if a device does not
support the required concurrency. We will need to add in special cases
for past boards which met AVL prior to those requirements being added.

For (2), This test will never be used to verify actual support. That
role will be taken by a more comprehensive concurrency validation test.

BUG=b:227217920
TEST=./fast_build.sh -T
tast run ${DUT} wifi.ConcurrencyCaps
tast run ${DUT} wifi.Caps

Change-Id: I7d02dc09e038918106f45276e03e068e0562aecb