Tast test: Fix flaky ui.NotificationClosePerf
ui.NotificationClosePerf is flaky when we add ARC notification. This is
due to the fact that when we interact with the ARC testing app,
sometimes we loose focus of the message center and it closes by itself,
causing failure while trying to interact with message center UI. To fix
this, we will now open the message center at every iteration after all
ARC notifications are created.

BUG=chromium:1302447
TEST=tast run <IP> ui.NotificationClosePerf

Change-Id: I596079168729ea6016fb2ef68812c6197d1f9d4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3560801
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Tested-by: Alex Newcomer <newcomer@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Andre Le <leandre@chromium.org>